Tesla takes damage

+++ Assets/Scripts/Entity.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Entity : MonoBehaviour {
 
-	protected int _health;
+	protected float _health;
 	protected int _strength;
